refactor(renamed files): changed to kebab-case
