+++ src/components/delete-icon/index.tsx
