--- src/pages/index.tsx
@@ -1,5 +1,0 @@
-import Home from './Home';
-import Game from './Game';
-import Scoreboard from './Scoreboard';
-
-export { Home, Game, Scoreboard };
